parser-server/requirements.txt
grpcio==1.37.1
grpcio-tools==1.37.1
TexSoup==0.3.1
pyinstaller==4.3